requirements.txt
biopython==1.79
click==8.1.3
daiquiri==3.2.1
networkx==2.7.1
numpy==1.21.0
pandas==1.4.2
progressbar2==4.0.0
requests==2.28.1
scipy==1.7.3
setuptools==61.2.0
statsmodels==0.13.2
pypdl==1.0.5